Do not auto remove direct-opened attachments
Previously, Trojita tried to wait for a while and remove the temporary
file on disk 10 seconds after the application has been launched. That
was a possible race, and worse, some applications attempt to be smart
and detect this file removal and somehow handle that. An example is
Gwenview which auto-switches to a next file in the directory. Another
example is Ark which won't let you extract/preview additional files from
the archive once the downloaded file has been removed.

It's 2017, and it seems that we can now mostly rely on a system behaving
"sanely" and cleaning /tmp automatically. That's probably better than a
random ad-hoc timeout of 10 seconds which we used to have in Trojita.

Another bonus is that this fixes an interoperability problem with Adobe
Reader on Windows. That app complained when the file to be openened was
locked by another application -- Trojita in this case.

BUG: 344606
Change-Id: Ic150c6d2976ac1a414a6b34a41008aed104c7098
CCBUG: 316981