Use ‘resend’ instead of ‘bounce’ in GUI and source
Trojitá supports ‘Resending’ messages, where a message is taken,
some Resent-headers are added at the top, and then sent otherwise
unchanged. This was described as ‘Bouncing’ in the source and
(partly) in the GUI. Bouncing is more often used to denote sending
a message to the sender signalling non-reception; such
functionality is also included in some MUAs to fake non-reception.

To avoid confusion, all terminology has been migrated from
‘bounce’ to ‘resend’, both in strings shown in the GUI and names
used in the source code, including the name of the as of yet
non-existent icon. This includes the correction of an incorrect
naming of the resending action as ‘Edit as New E-Mail Message’.

Change-Id: If8abdafb7a7c867d0d70f118561fdb7270bd96ce