restructure menu bar: use app/account/mailbox/message division
Currently, Trojitá's menu bar consists of four main menus: IMAP,
View, Mailbox, Help. Apart from View. These names do not cover
many of the actions they contain. This patch attempts to create
a menu structure where the actions are grouped in a more
meaningful way.

Namely, the IMAP and Help menus are removed, and Application,
Account, Message, and Plugins menus are added. The actions in
the old structure are moved into these and the pre-existing
Mailbox menu so as to arrive at a consistent whole.

While the discoverability of actions is improved in this new
structure, it could still be made more consistent. For example,
the Settings action effectively covers settings that are related to
the Application, Account, and Plugin menus. Also, the debugging
functionality is kept grouped under Application, but could
conceivably be distributed over the Application, Account, and
Message menus.

Many actions in the context menus do not appear in the menu
bar menus, while some would also fit there. This has not been
tackled in this patch. The focus here is kept on the restructuring.

Change-Id: Idce348d9e4350867fc888701b7ea51f2006715af
Signed-off-by: Jan Kundrát <jkt@kde.org> (additional reshuffling)