enable Flag, (Not) Junk, and Archive shortcuts in message list
The actions for Flagging, Marking as (Not) Junk, and Archiving were
not added to msgListWidget->tree. So if the toolbar was not present
or the context menu not open, the corresponding shortcuts would
not work. (Instead the message list ‘cursor’ would move to the next
message whose subject starts with the shortcut letter.)

The actions have now been added, as was already the case for
Delete, Read, etc.

Change-Id: Iaf10bddf49bc309d641a27299f96f36a514c620f