Make signature separator regexp stricter to avoid wrong positives
The signature separator regexp includes a pattern for the footer
separator used by default by Mailman (until recently [*]), captured
as a ‘line’ of 45 or more underscores. Unsurprisingly there are
messages that include lines of underscores not meant to be
signature separators. For me those had quite a bit more than
45 underscores. So this patch introduces an upper bound (of 55)
in the pattern to try and make a better distinction. (Of course it
would be best to know what Mailman used as a default, but I
have not been able to determine this; in my example mails, it
was 47.)

[*] https://bugs.launchpad.net/mailman/+bug/266269

Change-Id: I83b40811d98efaf23a6a84d469b0d809a2340a90