Transition from QRegExp to QRegularExpression
Qt5 introduced a new regex class, QRegularExpression, that is
supposed to be better. The existing class QRegExp is said to be
unmaintained, buggy, and less featurefull. This patch transitions
Trojita from QRegExp to QRegularExpression. This is done by
initial ‘literal translation’ without making use of any new
functionality. Then, if a test failed, more extensive rewriting
was done to fix the literal translation, also making use of the
extra functionality.

This patch future-proofs Trojitá and paves the way for making use
of the extra functionality QRegularExpression has to offer
(e.g., multiline and multiple matches) in existing and new code.

This patch does not transition the test suite, as we need it to
work as before to test this patch.

Change-Id: Ib8472934274fb97f8c67f44e969a5c75ac434fec