IMAP: fall back to binary part rendering when BODYSTRUCTURE parsing fails
We are relying on the server-side parsing of MIME messages. I just got a
report from "leto" on IRC which essentially boils down to the fact that
GMail tends to send garbled BODYSTRUCTURE of a message/rfc822 which is
enclosed in a multipart/report. The message/rfc822 has NIL for both body
and envelope.

We've been handling a NIL envelopes since e57849112 ("It is not allowed
by the RFC3501, but gmail occasionally sends that", does it ring a bell,
perhaps?)." The failure on pwned bodies was hard, however.

This patch treats a missing body of a message/rfc822 as if the whole
container wasn't a message/rfc822, but just a random binary MIME part.
It is important to reset the MIME type so that the upper layers of
Trojita do not assume (and assert on) existence of the .HEADER and .BODY
sub-indexes in the part tree.

Change-Id: I006d467fb38510901743b6608e03d9c664dceb17