crypto: do not assert-crash on disappearing messages
This is quite unlikely to be hit in real world (I've only hit this once,
and there's one other bug report about this), but if we get reconnected
(and mailboxes trashed) at just the wrong moment when a message would
normally be populated based on remote server's responses, the code would
assert crash here.

This is mainly a blind fix, I have not put much effort into trying to
come up with a solid autotest for this.

BUG: 392800
Change-Id: I3b7539bebc5de0b5c10c4c05d683cd746aca2402