GUI: detect some calendaring events
So, I started getting many text/calendar items from
$corporate_environment. Let's try if Trojita's experience when dealing
with this can be improved a little bit.

The first step is determining the kind of the calendar object from MIME.
This is standardized [1] and provided by IMAP's usual BODYSTRUCTURE. I
was thinking that the UX could be made better by directly showing
whether this is:

- an invitation which typically requests some action or targets "me",
- some ad-hoc event notification,
- a reply to an invitation, etc.

This can be done even without parsing ical files at all, which is nice.
What I would like to see in future is some basic UI rendering of the
calendar object, integration with an external calendar system (hi
vdirsyncer+khal, etc), and a way of sending replies
(confirmation/rejection) when I'm invited to a meeting. That's for
future work, though.

[1] https://tools.ietf.org/html/rfc5546#section-3.2

Change-Id: Icfd90ad3dad3d5d6ae5a2c7dce1cad8f44b89ee2