Fix colors of quote bars when using user-provided CSS
The code tried to be smart when reading the user-provided CSS from the
disk, only re-interpreting it if the file's timestamp changed. I think
it's probably not critical to optimize that, but that commit
(af2c4f34b5) was still correct.

Stuff changed when I did 5d3696a61, ddbbcf057. With these commits and
when using a user-provided CSS, the message viewer would no longer use
correct colors for the quote indicator because that call to
QString::arg() for replacing the %1 and %2 placeholders was undone by
reinitialization of that variable from the defaultStylesheet.

Fix this by simplifying the logic at the cost of some extra disk IO.
Hey, we're invoking a full blown HTML viewer, let's not optimize stuff
that's not in a hotspot.

Change-Id: Ic399b49c9f6f73fb3da9bff04ddb0d7646e317cf