Move from Gerrit to KDEs Gitlab for merge requests
The certificate on gerrit.vesnicky.cesnet.cz expired
October 1, 2019, and KDE doesnt own or control the
certificate, so the infrastructure team cannot fix it.

Seeing as the rest of KDE is using Gitlab, it makes sense
that Trojita also starts using it, as it might be less of
a hurdle for drive by contributors to contribute.

BUG: 412841