Replace deprecated std::bind2nd (and std:binary_function)
std::bind2nd has been deprecated since C++11 and removed in C++17.
The same holds for std::binary_function, which is used in support of
std:bind2nd. A good way to replace the former is to use lambdas,
which is what is done in this commit. It is possible to get rid of
the latter by using plain functions. This set of changes makes the
code future-proof and removes a few warnings when compiling.

Signed-off-by: Jan Kundrát <jkt@kde.org>