Add override specifiers as suggested by cppcheck
A run of cppcheck on the codebase resulted, among other things,
in a long list of suggestions about adding override specifiers
when overriding ancestor class methods. This commit applies those
(and removes the then implied virtual) to get the possible (minor)
advantage associated to that and get a better overview of other
suggestions made by cppcheck.