call pathToPart() on Trojita-specific URLs only
The code was feeding all URLs through the
MsgPartNetAccessManager::pathToPart(). This had a side effect that those
URLs whose paths looked like a path from a valid Trojita-specific URL
would try to walk the part tree. This in itself is not a problem
(pathToPart() is public API and therefore it should not make any
assumptions about URL validity), but as the previous commit shows,
sometimes mistakes happen. The funny URL looks like this one:

 http://example.org/2022/06/11/foo.png

Since pathToPart() only makes sense on Trojita-specific URLs, call that
just in that case.